+++ src/main/java/timkodiert/budgetBook/view/ControllerFactory.java
@@ -5,8 +5,10 @@ import java.util.Map;
 import java.util.function.Supplier;
 import javax.inject.Inject;
 
-import timkodiert.budgetBook.view.fixedExpenses.FixedExpensesManageView;
-import timkodiert.budgetBook.view.uniqueExpenses.UniqueExpensesManageView;
+import timkodiert.budgetBook.view.fixed_expense.FixedExpenseDetailView;
+import timkodiert.budgetBook.view.fixed_expense.FixedExpensesManageView;
+import timkodiert.budgetBook.view.unique_expense.UniqueExpenseDetailView;
+import timkodiert.budgetBook.view.unique_expense.UniqueExpensesManageView;
 
 public class ControllerFactory {
 
@@ -21,11 +23,15 @@ public class ControllerFactory {
     }
 
     private void registerController() {
+        // Daten-/Steuerungsansichten
         viewControllerMap.put(AnnualOverviewView.class, viewComponent::getAnnualOverviewView);
         viewControllerMap.put(MonthlyOverview.class, viewComponent::getMonthlyOverview);
-        viewControllerMap.put(FixedExpensesManageView.class, viewComponent::getFixedExpenseManageView);
-        viewControllerMap.put(UniqueExpensesManageView.class, viewComponent::getUniqueExpensesManageView);
         viewControllerMap.put(ImportView.class, viewComponent::getImportView);
+        // MDV / Stammdaten
+        viewControllerMap.put(FixedExpensesManageView.class, viewComponent::getFixedExpenseManageView);
+        viewControllerMap.put(FixedExpenseDetailView.class, viewComponent::getEditExpenseView);
+        viewControllerMap.put(UniqueExpensesManageView.class, viewComponent::getUniqueExpensesManageView);
+        viewControllerMap.put(UniqueExpenseDetailView.class, viewComponent::getUniqueExpenseDetailView);
     }
 
     public View create(Class<?> viewControllerClass) {
