+++ src/main/java/timkodiert/budgetBook/view/fixed_expense/FixedExpenseDetailView.java
@@ -1,4 +1,4 @@
-package timkodiert.budgetBook.view.fixedExpenses;
+package timkodiert.budgetBook.view.fixed_expense;
 
 import java.net.URL;
 import java.util.List;
@@ -40,7 +40,7 @@ import timkodiert.budgetBook.util.CategoryTreeHelper;
 import timkodiert.budgetBook.util.DoubleCurrencyStringConverter;
 import timkodiert.budgetBook.util.EntityManager;
 import timkodiert.budgetBook.util.StageBuilder;
-import timkodiert.budgetBook.view.baseViews.EntityBaseDetailView;
+import timkodiert.budgetBook.view.mdv_base.EntityBaseDetailView;
 
 import static timkodiert.budgetBook.util.CategoryTreeHelper.from;
 
@@ -134,11 +134,6 @@ public class FixedExpenseDetailView extends EntityBaseDetailView<FixedExpense> i
             setEntity(null);
             onUpdate.run();
         }
-    }
-
-    @Override
-    public String getFxmlLocation() {
-        return "/fxml/EditExpense.fxml";
     }
 
     @Override
